Modified dependency 2 files : notice.sql + reservation.sql + user_company.sql + user_event.sql for create & insert Dependency 2 update for : install_db.sql + insert_db.sql + delete_db.sql

+++ sql_scripts/create/reservation.sql
@@ -1,6 +1,6 @@
 DROP TABLE IF EXISTS reservation;
 CREATE TABLE reservation (
-    id_reservation BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id_reservation INT AUTO_INCREMENT PRIMARY KEY,
     id_user INT,
     number_room INT,
     reference_promo INT,
